+++ lib/hexapdf/layout/box.rb
@@ -144,7 +144,7 @@ module HexaPDF
       attr_reader :properties
 
       # :call-seq:
-      #    Box.new(width: 0, height: 0, style: nil, properties: {}) {|canv, box| block} -> box
+      #    Box.new(width: 0, height: 0, style: nil, properties: nil) {|canv, box| block} -> box
       #
       # Creates a new Box object with the given width and height that uses the provided block when
       # it is asked to draw itself on a canvas (see #draw).
@@ -152,11 +152,11 @@ module HexaPDF
       # Since the final location of the box is not known beforehand, the drawing operations inside
       # the block should draw inside the rectangle (0, 0, content_width, content_height) - note that
       # the width and height of the box may not be known beforehand.
-      def initialize(width: 0, height: 0, style: nil, properties: {}, &block)
+      def initialize(width: 0, height: 0, style: nil, properties: nil, &block)
         @width = @initial_width = width
         @height = @initial_height = height
         @style = Style.create(style)
-        @properties = properties
+        @properties = properties || {}
         @draw_block = block
         @fit_successful = false
         @split_box = false